Add optional OSC test combinatorial run
Test OSC functions:
  hyperlink
  title

With a variety of terminal emulators and shells. The test
results only check to see if the execution succeeded.
Since the "output" of these tests rely on viewing the
terminal emulator It's up to the person running these tests
to confirm that they can observe the output and the output
is as-expected.

Because the results do not fit within the typical test
run (i.e. in a CI pipeline) these tests are marked with #[ignore]
and must be run manually with:

  cargo test -- --ignored --show-output