Enable wrapping of OSC zero-width sequences
Depending on the context users of the OSC API may need to
cause the OSC sequences to be wrapped in strings indicating
the start and end of zero-width sequences.

Different shells use different sequences in different
contexts to denote the beginning and end of zero-width
sequences. In the context of forming the command prompt
shells tend to use plain ASCII sequences of two characters.

Bash, for example, translates the ASCII text "\[" when found
in prompt strings to the byte 0x01 (aka SOH aka Start Of
Heading aka Ctrl-A) to indicate the start of a zero-width
sequence. Similarly Bash translates "\]" to 0x02 (aka STX
aka Start of TeXt aka Ctrl-B).

Meanwhile Zsh and TCSH use the ASCII text "%{" and "}%"
in prompt strings.

Given this variety of possible strings a caller may wish to
produce we offer two often-useful simple enums:

  None (the default)
  Wrapping::CtrlACtrlB

and one which is maximally adaptable:

  Wrapping::Str(begin, end)

Rather than take a parameter to only ::hyperlink(), allow
callers to set the wrapper directly on AnsiGenericString.
This allows wrapping of zero-width sequences to later be
extended to color sequences with zero additional changes
to the API.

If additional wrapping outside the context of zero width
sequences ever becomes useful the Wrapping type is also
ready to be used there.